tjpg decode the jpeg block by block

+++ src/jpeg_decoder/jpeg_decode.h
@@ -4,7 +4,7 @@
 #include "tjpgdcnf.h"
 #include "tjpgd.h"
 uint16_t jd_input(JDEC* jdec, uint8_t* buf, uint16_t len);
-int jd_output(JDEC* jdec, void* bitmap, JRECT* jrect);
+uint16_t jd_output(JDEC* jdec, void* bitmap, JRECT* jrect);
 void tjpg_test();
 
 #endif
